Adding support for jailer
Adds support for firecracker's jailer. This allows for more secure
virtualization through creation of namespaces and cgroups. Jailer is on
by default when creating a new machine. However, this can be turned off
by setting DisableJailer to true in the configuration.

rebasing to master

Adding support for jailer

Adds support for firecracker's jailer. This allows for more secure
virtualization through creation of namespaces and cgroups. The jailer
can be turned on by the setting the EnableJailer field in the config.
This also implements a very naive mapper and will require update to
support a smarter one.

moving inlined handlers to be singletons

previous commit had handlers as closures but based on feedback, it was
decided to move those to a singleton instead.

add dev mapper strategy to machine's config

allocating exitCh and adding root disable env var

Tests that require root can now be disabled by setting the
DISABLE_ROOT_TESTS flag.