Adding custom templates for go-swagger
This gets rid of the need for mockgen and uses go-swagger to instead
generate a mock client. In addition, this change also gets rid of the
wrapper client and instead uses the client.Firecracker instead. Also,
the change also generates an interface in the operations package and
also needed some updating to the default go-swagger template to use the
interface instead of types directly.