machine: make Wait() safe for multiple calls
Prior to this change, multiple calls to Wait() could result in the error
being lost, and for subsequent callers to wait until their context was
canceled rather than waiting for the VMM to exit.  This change makes it
such that multiple callers of Wait() will all block for the VMM exit (or
their individual context cancellation) and will receive the same error
(for non-canceled contexts).

Signed-off-by: Samuel Karp <skarp@amazon.com>