network: set broad permissions
Some of the tests in this package are expected to run as root in order
to perform privileged operations.  These tests may also create files,
but if they do so without setting the w bit for the other group, the
files will not be deletable after the test by a normal user.  This
commit sets the writable bit on the file permissions only so that the
files can be deleted by non-root users after the tests complete.

Signed-off-by: Samuel Karp <skarp@amazon.com>