Fixes hanging when using FifoLogWriter
This change fixes the issue with the FifoLogWriter during the
captureFifoToFile call. During os.OpenFile, the operation would block
indefinitely waiting for firecracker to write to the fifo since that
logic was moved before firecracker could do so. This fix will now call
syscall.Open with the syscall.O_NONBLOCK flag instead.

Signed-off-by: xibz <impactbchang@gmail.com>