Do not overwrite handlers in LoadSnapshot
As LoadSnapshot requires only a subset of the default handlers, it was
implemented by assigning the FcInit handlers to a smaller list of
handlers. However, this meant that it would overwrite the FcInit
handlers given by the jailer. This PR fixes this behavior by just
removing the unnecessary handlers, instead of overwriting them entirely.

Signed-off-by: David Son <davbson@amazon.com>